Update code
This code i written for reminder application. This code written in application server class.

Whenever the cache item expires, we get a callback and the item is gone from the cache. So, we no longer get any callback in future. In order to have a continuous supply of callback, we need to store an item in cache again upon expiration. based on this i have written this code to send email and sms to the users from my application